Handle user authorization

+++ src/controllers/routines.ts
@@ -2,11 +2,14 @@ import { RequestHandler } from "express";
 import RoutineModel from "../models/routine"
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 
 export const getRoutines: RequestHandler = async (req, res, next) => {
+    const authenticatedUserId = req.session.userId;
     try {
-        const routines = await RoutineModel.find().exec();
+        assertIsDefined(authenticatedUserId);
+        const routines = await RoutineModel.find({userId: authenticatedUserId}).exec();
         res.status(200).json(routines);
     } catch (error) {
         next(error);
@@ -15,8 +18,12 @@ export const getRoutines: RequestHandler = async (req, res, next) => {
 
 export const getRoutine: RequestHandler = async (req, res, next) => {
     const id = req.params.routineId;
+    const authenticatedUserId = req.session.userId;
+
 
     try {
+        assertIsDefined(authenticatedUserId);
+
         if (!mongoose.isValidObjectId(id)) {
             throw createHttpError(400, "invalide Routine Id");
         }
@@ -25,6 +32,10 @@ export const getRoutine: RequestHandler = async (req, res, next) => {
 
         if (!routine) {
             throw createHttpError(404, 'routine not found');
+        }
+
+        if (!routine.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You can't access this note.");
         }
 
         res.status(200).json(routine);
@@ -41,13 +52,17 @@ interface CreateRoutineBody {
 export const createRoutine: RequestHandler<unknown, unknown, CreateRoutineBody, unknown> = async (req, res, next) => {
     const title = req.body.title;
     const text = req.body.text;
+    const authenticatedUserId = req.session.userId;
+
 
     try {
+        assertIsDefined(authenticatedUserId);
         if (!title) {
             throw createHttpError(400, "Routine must have a title.")
         }
 
         const newRoutine = await RoutineModel.create({
+            userId: authenticatedUserId,
             title: title,
             text: text,
         });
@@ -71,8 +86,12 @@ export const updateRoutine: RequestHandler<UpdateRoutineParams, unknown, UpdateR
     const newTitle = req.body.title;
     const newText = req.body.text;
     const id = req.params.routineId;
+    const authenticatedUserId = req.session.userId;
+
 
     try {
+        assertIsDefined(authenticatedUserId);
+
         if (!mongoose.isValidObjectId(id)) {
             throw createHttpError(400, 'bad Id')
         }
@@ -85,6 +104,10 @@ export const updateRoutine: RequestHandler<UpdateRoutineParams, unknown, UpdateR
 
         if(!routine) {
             throw createHttpError(400, "there's no such routine")
+        }
+
+        if (!routine.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You can't access this note.");
         }
 
         routine.title = newTitle;
@@ -101,8 +124,12 @@ export const updateRoutine: RequestHandler<UpdateRoutineParams, unknown, UpdateR
 
 export const deleteRoutine: RequestHandler = async(req, res, next) => {
     const id = req.params.routineId;
+    const authenticatedUserId = req.session.userId;
+
 
     try {
+        assertIsDefined(authenticatedUserId);
+
         if (!mongoose.isValidObjectId(id)) {
             throw createHttpError(400, 'bad Id')
         }
@@ -111,6 +138,10 @@ export const deleteRoutine: RequestHandler = async(req, res, next) => {
 
         if (!routine) {
             throw createHttpError(404, "there's no such routine")
+        }
+
+        if (!routine.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You can't access this note.");
         }
 
         await routine.deleteOne();
